requirements.txt
asgiref==3.5.0
Django==2.2.27
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.0.4
inflection==0.5.1
psycopg2==2.8.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2